Removing extra test class, setting the main game to reflect changes. Libgdx updated to x.x.5, making minor behavioral "fixes"
